Removed obsolete wraparound code
Wraparound is now handled by the Contour-engine by shifting the matrix
indices so that contouring starts at the position from where coordinates
will change in a consistent manner. The exact position will depend on
the start position of the global data (-180 for EC, 0 for GFS querydata),
and the selected world view (Atlantic, Pacific, or some other central longitude).