Fixed more double vs float rounding issues
Isoline and isoband limits must be stored as floats instead of doubles, otherwise for example
isolines which do not compare equal as double and float will vanish since the vertices will
be marked as ghost elements to be removed from isobands to get the final isoline. The altenative
would have been to allow an epsilon in the comparison, but consistent use of float is cleaner.